We have 8 health statuses
there by eveyr time we take damage it is useful to
to just how many of those 8 health statuses we have lost.
to that end i divide the starting health of 100/8. which yields 12.5
so if you take a damage of 10. due to behavior of convent.toint.
converts
10/12.5 into nearest integer of 1. so you lose 1 health status level.

also if you enter dying stage it does not make sense for you remain
alive. you should just die (We dont have a timer at this time.) or else
this could be soemthing that took say 10 seconds.

other changes. re order health statuss.
so higher index means you are higher health level (this make the math
easier to read). also mde the colors green and red a bit different. to
make them different.

see that now strong is super strong are different shades of green.